chore(level_detection): Make log level detection configurable (#14784)
