+++ pkg/distributor/limits.go
@@ -24,6 +24,7 @@ type Limits interface {
 	IncrementDuplicateTimestamps(userID string) bool
 	DiscoverServiceName(userID string) []string
 	DiscoverLogLevels(userID string) bool
+	LogLevelFields(userID string) []string
 
 	ShardStreams(userID string) shardstreams.Config
 	IngestionRateStrategy() string
