+++ pkg/validation/limits_test.go
@@ -216,6 +216,7 @@ ruler_remote_write_headers:
 			exp: Limits{
 				RulerRemoteWriteHeaders: OverwriteMarshalingStringMap{map[string]string{"foo": "bar"}},
 				DiscoverServiceName:     []string{},
+				LogLevelFields:          []string{},
 
 				// Rest from new defaults
 				StreamRetention: []StreamRetention{
@@ -234,7 +235,7 @@ ruler_remote_write_headers:
 `,
 			exp: Limits{
 				DiscoverServiceName: []string{},
-
+				LogLevelFields:      []string{},
 				// Rest from new defaults
 				StreamRetention: []StreamRetention{
 					{
@@ -254,6 +255,7 @@ retention_stream:
 `,
 			exp: Limits{
 				DiscoverServiceName: []string{},
+				LogLevelFields:      []string{},
 				StreamRetention: []StreamRetention{
 					{
 						Period:   model.Duration(24 * time.Hour),
@@ -274,6 +276,7 @@ reject_old_samples: true
 			exp: Limits{
 				RejectOldSamples:    true,
 				DiscoverServiceName: []string{},
+				LogLevelFields:      []string{},
 
 				// Rest from new defaults
 				RulerRemoteWriteHeaders: OverwriteMarshalingStringMap{map[string]string{"a": "b"}},
@@ -293,7 +296,9 @@ query_timeout: 5m
 `,
 			exp: Limits{
 				DiscoverServiceName: []string{},
-				QueryTimeout:        model.Duration(5 * time.Minute),
+				LogLevelFields:      []string{},
+
+				QueryTimeout: model.Duration(5 * time.Minute),
 
 				// Rest from new defaults.
 				RulerRemoteWriteHeaders: OverwriteMarshalingStringMap{map[string]string{"a": "b"}},
